Implemented TOX test environment added docs and unit tests
This is a huge commit where only test environment was meant to be added,
but original functions needed to be changed because originals were hard
to test and would require mocking a file. Since that would add
complexity of otherwise simple plugin rewrite was easier.

Adding also to travis.yml and renaming package to mlx.warnings instead
of probably frequently used ci.