CodeClimate configuration needs to be adjusted
That is because they are not checking the issued marked as invalid and
we have lower score because of that. Current adjustment makes the
duplication detection limit higher, but that will also mean that it will
catch less errors. Once the issue on their side is fixed, we can again
move back to defaults.