Adding stream parsing to plugin
There are cases where you would just pipe stderr or stdout to the
plugin. To avoid additional pipe work the plugin can run the command
instead of you.

Currently bad part is that it runs it in a subprocess and if you have a
long command the progress is not reported- until later.