Quick and dirty implementation of Coverity checker
Pulled in Sphinx Coverity plugin, so that functions for querying
Coverity Connect Server could be used. The configuration is read from
.env file (or environment variables) while command to actually invoke
the plugin is

```
mlx-warnings --coverity bla
```

That is because we want to actually bypass argparse and required
logfile/command argument. It might be useful to use command here (like
`cov-commit-defects`) and then query the DB for results.